rename styleUtils file
