+++ src/modules/map/components/olMap/handler/geolocation/styleUtils.js
